refactor: switch formik to rhf for auth forms
