+++ src/lib/components/auth/Login/constants.ts
@@ -1,0 +1,14 @@
+import * as yup from "yup";
+
+import type { LoginFormType } from "./types";
+
+export const loginFormValidationSchema: yup.SchemaOf<LoginFormType> =
+  yup.object({
+    email: yup.string().email().required().defined(),
+    password: yup.string().required().defined(),
+  });
+
+export const initialValues: LoginFormType = {
+  email: "",
+  password: "",
+};
