+++ src/lib/components/auth/Login/types.ts
@@ -1,0 +1,8 @@
+export type LoginFormType = {
+  email: string;
+  password: string;
+};
+
+export type LoginProps = {
+  handleSwitchToRegister: () => void;
+};
