+++ src/lib/components/auth/Register/index.tsx
@@ -1,0 +1,140 @@
+import {
+  Button,
+  Grid,
+  Heading,
+  ModalBody,
+  ModalContent,
+  ModalFooter,
+  ModalHeader,
+  useColorModeValue,
+  useToast,
+} from "@chakra-ui/react";
+import { yupResolver } from "@hookform/resolvers/yup";
+import { useState } from "react";
+import { useForm } from "react-hook-form";
+
+import SignInProviders from "lib/components/auth/SignInProviders";
+import { contraBoxStyle } from "lib/components/auth/style";
+import SpokerInput from "lib/components/shared/SpokerInput";
+import { registerUserWithEmailAndPassword } from "lib/services/firebase";
+import { trackEvent } from "lib/utils/trackEvent";
+
+import { initialValues, registerFormValidationSchema } from "./constants";
+import type { RegisterFormType, RegisterProps } from "./types";
+
+const Register = ({ handleSwitchToLogin }: RegisterProps) => {
+  const toast = useToast();
+  const [isLoading, setIsLoading] = useState<boolean>(false);
+  const borderColor = useColorModeValue("#18191F", "#FFFFFF");
+
+  const {
+    register,
+    getValues,
+    handleSubmit,
+    formState: { isDirty, isValid, errors },
+  } = useForm<RegisterFormType>({
+    defaultValues: initialValues,
+    mode: "onChange",
+    resolver: yupResolver(registerFormValidationSchema),
+  });
+
+  const processRegister = async () => {
+    setIsLoading(true);
+    const values = getValues();
+    await registerUserWithEmailAndPassword(
+      values.email,
+      values.password,
+      values.name
+    )
+      .then((user) => {
+        setIsLoading(false);
+        trackEvent("New User Register", "registration");
+        toast({
+          title: "Registration Successful",
+          position: "top",
+          status: "success",
+          isClosable: true,
+        });
+        toast({
+          title: `A verification email is sent to ${user?.email}`,
+          description:
+            "Before you can use any features, please verify your email first.",
+          position: "top",
+          status: "warning",
+          isClosable: true,
+        });
+      })
+      .catch((err) => {
+        setIsLoading(false);
+        toast({
+          description: err.message,
+          position: "top",
+          status: "error",
+          isClosable: true,
+        });
+      });
+  };
+
+  return (
+    <ModalContent
+      as="form"
+      {...contraBoxStyle(borderColor)}
+      onSubmit={handleSubmit(processRegister)}
+    >
+      <ModalHeader>
+        <Heading bgGradient="linear(to-br, teal.200, green.400)" bgClip="text">
+          Register
+        </Heading>
+      </ModalHeader>
+
+      <ModalBody>
+        <Grid gap={4}>
+          <SpokerInput
+            label="name"
+            {...register("name")}
+            isInvalid={!!errors.name?.message}
+            errorText={errors.name?.message}
+            placeholder="What's your name?"
+          />
+          <SpokerInput
+            label="email"
+            {...register("email")}
+            isInvalid={!!errors.email?.message}
+            errorText={errors.email?.message}
+            placeholder="Your favorite email"
+          />
+          <SpokerInput
+            label="password"
+            {...register("password")}
+            isInvalid={!!errors.password?.message}
+            errorText={errors.password?.message}
+            placeholder="Your secret phrase"
+            type="password"
+          />
+
+          <SignInProviders />
+        </Grid>
+      </ModalBody>
+
+      <ModalFooter gridGap={2}>
+        <Button
+          variant="ghost"
+          fontWeight="normal"
+          onClick={handleSwitchToLogin}
+        >
+          Sign In
+        </Button>
+        <Button
+          type="submit"
+          disabled={!isDirty || !isValid}
+          isLoading={isLoading}
+          colorScheme="teal"
+        >
+          Sign Up
+        </Button>
+      </ModalFooter>
+    </ModalContent>
+  );
+};
+
+export default Register;
