+++ src/lib/components/auth/Register/types.ts
@@ -1,0 +1,9 @@
+export type RegisterFormType = {
+  name: string;
+  email: string;
+  password: string;
+};
+
+export type RegisterProps = {
+  handleSwitchToLogin: () => void;
+};
